requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
Pillow>=10.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
